Refactor hosted semantic backend: streamline catalog handling, centralize hosted-specific logic, and improve modularity with `hosted_catalog` introduction.
