+++ packages/dex-core/src/exmergo_dex_core/explore/semantic/hosted.py
@@ -18,34 +18,24 @@ from __future__ import annotations
 import json
 import os
 import re
-import time
 from dataclasses import replace
 from typing import Any
 
-from ... import envelope as env
 from ... import metricflow_dialect
 from ...config import QueryLimits
-from ...semantic_catalog import SemanticCatalogView, column_reference
+from ...semantic_catalog import SemanticCatalogView
 from ..results import SemanticQueryResult, SemanticValuesResult
 from . import (
     DIMENSIONS_PER_QUERYABLE_PATH,
     EXECUTION_VENDOR,
-    DimensionInfo,
-    EntityInfo,
-    EntityRole,
-    MeasureInfo,
-    MetricComposition,
+    BackendDescriptor,
     MetricInfo,
     SemanticBackendError,
     SemanticCatalog,
-    SemanticModelInfo,
     SemanticQuery,
     SemanticQueryRefusedError,
     ValuesRequest,
-    cap_columnar,
     cost_posture,
-    derive_entity_type,
-    merge_element_fields,
     merge_pii_meta,
     queryable_grains,
     requested_dimension_refs,
@@ -56,185 +46,19 @@ from . import (
     validate_grain,
     values_reach_note,
 )
-
-# The very explicit line the founder asked for: wherever a reader or agent might
-# expect the cost guard, say plainly that it does not apply on the hosted path.
-_HOSTED_COST_WARNING = (
-    "cost guard unavailable on the hosted semantic layer: dbt Cloud owns the "
-    "warehouse connection and executes this query server-side, so dex applies no "
-    "cost estimate or ceiling here. Spend is governed by the dbt Cloud "
-    "environment's own limits, not by dex."
+from .hosted_catalog import (
+    HOSTED_CATALOG_FIELDS,
+    HOSTED_CATALOG_GAPS,
+    HOSTED_COST_WARNING,
+    decode_hosted_catalog,
+    hosted_grains,
 )
-
-# What a hosted catalog cannot answer, declared per element kind rather than left
-# to be inferred from a missing key. Measured by introspecting the live schema:
-# `Entity` has no `label` field at all, `SemanticModel` carries only `name`, and
-# `Measure` carries no words. A silent absence would read as a dbt project that
-# documented none of it, which for a well-documented project is the opposite of
-# the truth, and a note alone is the part of a payload a caller truncates first.
-_HOSTED_CATALOG_GAPS: dict[str, list[str]] = {
-    "semantic_models": [
-        "label",
-        "description",
-        "model_ref",
-        "agg_time_dimension",
-        "primary_entity",
-        "relation",
-    ],
-    "entities": ["label"],
-    "measures": ["label", "description"],
-}
-
-# A hosted catalog cannot reach the physical layer at all: no relation on a
-# semantic model, so no address for any element's column either. Worth prose as
-# well as the declaration, because the question it blocks ("which table is behind
-# this metric") is one a caller arrives with rather than discovers.
-_HOSTED_RELATION_NOTE = (
-    "hosted list: the dbt Cloud Semantic Layer API exposes no physical relation "
-    "on a semantic model (its SemanticModel type carries only a name), so every "
-    "element here names the column behind it and none of them can say which table "
-    "that column is in. List with --local for the relations, which is also what "
-    "connects a metric to the objects `explore map` and `explore profile` describe"
-)
-
-# The one gap worth prose as well, because it is the field a caller is most
-# likely to go looking for and there is somewhere better to get it.
-_HOSTED_ENTITY_LABEL_NOTE = (
-    "hosted list: the dbt Cloud Semantic Layer API exposes no label on entities "
-    "(its Entity type has no such field), so an entity here carries a "
-    "description at most; list with --local to read the labels the dbt project "
-    "declares"
-)
-
-# A hosted catalog is reached metric by metric, so an element no metric touches is
-# not in it. Said only when it can actually bite, which is any layer with metrics.
-_HOSTED_REACH_NOTE = (
-    "hosted list: every element is reached through a metric, so a measure, "
-    "entity declaration or semantic model that no metric draws on is absent; "
-    "list with --local for the layer as the project declares it"
-)
+from .hosted_transport import await_result, post_graphql, shape_json_result
 
 _MISSING_EXTRA = (
     "the hosted semantic-layer backend needs the [semantic-api] extra: "
     "pip install 'exmergo-dex-core[semantic-api]'"
 )
-
-
-def _model_name(value: Any) -> str | None:
-    """A ``semanticModel { name }`` selection reduced to the name it carries."""
-
-    return value.get("name") if isinstance(value, dict) else None
-
-
-def _bare_dimension(token: str | None, semantic_model: str | None) -> str | None:
-    """The bare dimension name behind a queryable token.
-
-    The API returns the token a query groups by (``agent__session__created_at``),
-    never the declaration behind it, so the bare name is read off the token: the
-    qualifier is a chain of entity names and ``__`` is the separator MetricFlow
-    itself uses, so what follows the last one is the dimension. Together with the
-    owning semantic model it identifies the declaration, which is what lets a
-    caller see that several paths reach one dimension.
-
-    Only where an owning model is known, which is what keeps dex's own synthesized
-    ``metric_time`` from being given a declaration it does not have.
-    """
-
-    if not token or not semantic_model:
-        return None
-    _, separator, tail = token.rpartition("__")
-    return tail if separator else token
-
-
-def _metric_info(payload: dict[str, Any], owners: list[str]) -> MetricInfo:
-    """One metric from the hosted catalog, composition included.
-
-    Composition is kept portable and MetricFlow's own vocabulary is kept apart:
-    a ratio's two sides and a derived metric's inputs mean the same thing in any
-    semantic layer, while a cumulative window or a grain-to-date only means
-    something here, so it travels under one declared vendor key.
-    """
-
-    params = payload.get("typeParams")
-    params = params if isinstance(params, dict) else {}
-
-    def named(value: Any) -> str | None:
-        return value.get("name") if isinstance(value, dict) else None
-
-    def window(value: Any) -> dict[str, Any] | None:
-        if not isinstance(value, dict):
-            return None
-        count, granularity = value.get("count"), value.get("granularity")
-        if count is None and granularity is None:
-            return None
-        return {"count": count, "granularity": granularity}
-
-    inputs = [
-        name
-        for name in (named(x) for x in params.get("inputMeasures") or [])
-        if name is not None
-    ]
-    input_metrics = [
-        name
-        for name in (named(x) for x in params.get("metrics") or [])
-        if name is not None
-    ]
-
-    vendor: dict[str, Any] = {}
-    if (cumulative := window(params.get("window"))) is not None:
-        vendor["window"] = cumulative
-    if params.get("grainToDate"):
-        vendor["grain_to_date"] = params["grainToDate"]
-    offsets = {
-        x["name"]: offset
-        for x in params.get("metrics") or []
-        if isinstance(x, dict)
-        and isinstance(x.get("name"), str)
-        and (offset := window(x.get("offsetWindow"))) is not None
-    }
-    if offsets:
-        vendor["offset_windows"] = offsets
-
-    if payload.get("requiresMetricTime"):
-        # Written only when true, so an absent key means false and a layer of
-        # dozens of metrics does not pay for a false on each of them.
-        vendor["requires_metric_time"] = True
-
-    # One name over several columns: each measure aggregates over its own time
-    # dimension, so a metric drawing on measures from two models has two.
-    time_axis = sorted(
-        {
-            measure["aggTimeDimension"]
-            for measure in payload.get("measures") or []
-            if isinstance(measure, dict) and measure.get("aggTimeDimension")
-        }
-    )
-
-    metric_filter = payload.get("filter")
-    return MetricInfo(
-        name=payload.get("name"),
-        type=(payload.get("type") or "").lower(),
-        label=payload.get("label"),
-        description=payload.get("description"),
-        dimensions=[d.get("name") for d in (payload.get("dimensions") or [])],
-        semantic_models=owners or None,
-        input_measures=inputs or None,
-        time_axis=time_axis or None,
-        queryable_granularities=_grains(payload),
-        composition=MetricComposition(
-            measure=named(params.get("measure")),
-            numerator=named(params.get("numerator")),
-            denominator=named(params.get("denominator")),
-            expr=params.get("expr"),
-            input_metrics=input_metrics or None,
-        ),
-        filter=(
-            metric_filter.get("whereSqlTemplate")
-            if isinstance(metric_filter, dict)
-            else None
-        ),
-        vendor_params=vendor or None,
-    )
 
 
 def _screening_notes(unknown: list[str], meta: dict[str, Any] | None) -> list[str]:
@@ -263,29 +87,6 @@ def _screening_notes(unknown: list[str], meta: dict[str, Any] | None) -> list[st
         "them. Mark a dimension with `meta: {pii: true}` in the dbt project to "
         "make the layer authoritative."
     ]
-
-
-_GRAIN_FIELDS = ("queryableTimeGranularities", "queryableGranularities")
-
-
-def _grains(payload: dict[str, Any]) -> list[str] | None:
-    """The grains the API reports for one metric or dimension, or None if it did
-    not answer.
-
-    Both fields are read and merged because they answer the same question in two
-    shapes: ``queryableGranularities`` is the standard enum, and
-    ``queryableTimeGranularities`` is the one that can also name a granularity the
-    project defined for itself. An **empty list is the answer** for a categorical
-    dimension, and it is the fact that stops a caller asking it for a month. A
-    payload carrying neither field is a deployment that was not asked, which is
-    not the same statement and must not read as one.
-    """
-
-    if not any(field in payload for field in _GRAIN_FIELDS):
-        return None
-    return metricflow_dialect.order_grains(
-        [value for field in _GRAIN_FIELDS for value in payload.get(field) or []]
-    )
 
 
 # Metric/dimension/entity names are identifiers; validating them keeps
@@ -325,12 +126,21 @@ class HostedDbtCloudBackend:
     # could price or cap. Everything the no-cost-guard posture needs follows from
     # this one declaration; see `cost_posture`.
     execution = EXECUTION_VENDOR
-    cost_guard_warning = _HOSTED_COST_WARNING
-    catalog_gaps = _HOSTED_CATALOG_GAPS
+    cost_guard_warning = HOSTED_COST_WARNING
+    catalog_gaps = HOSTED_CATALOG_GAPS
     # One row per token a query may group by, join-resolved by the API, including
     # dimensions reached through two joins. That is why this backend reports more
     # dimensions than a project read of the same layer.
     dimension_scope = DIMENSIONS_PER_QUERYABLE_PATH
+    descriptor = BackendDescriptor(
+        name=name,
+        vendor=vendor,
+        deployment=deployment,
+        execution=execution,
+        catalog_gaps=HOSTED_CATALOG_GAPS,
+        dimension_scope=dimension_scope,
+        cost_guard_warning=cost_guard_warning,
+    )
 
     _POLL_ATTEMPTS = 90
     _POLL_INTERVAL = 1.0
@@ -383,197 +193,26 @@ class HostedDbtCloudBackend:
     # ---- transport ---------------------------------------------------------
 
     def _post(self, query: str) -> dict[str, Any]:
-        import httpx
-
-        try:
-            with httpx.Client(timeout=self._timeout) as client:
-                resp = client.post(
-                    self._url,
-                    headers={
-                        "Authorization": f"Bearer {self._token}",
-                        "Content-Type": "application/json",
-                    },
-                    json={"query": query},
-                )
-        except httpx.HTTPError as exc:
-            raise SemanticBackendError(
-                f"could not reach the dbt Cloud Semantic Layer: {env.redact(str(exc))}"
-            ) from exc
-        if resp.status_code == 401 or resp.status_code == 403:
-            raise SemanticBackendError(
-                "dbt Cloud Semantic Layer rejected the token (HTTP "
-                f"{resp.status_code}); check DBT_SL_TOKEN is a current "
-                "'Semantic Layer Only' service token for this environment"
-            )
-        if resp.status_code != 200:
-            raise SemanticBackendError(
-                f"dbt Cloud Semantic Layer returned HTTP {resp.status_code}"
-            )
-        body = resp.json()
-        if body.get("errors"):
-            joined = "; ".join(str(e.get("message", e)) for e in body["errors"])
-            raise SemanticBackendError(f"semantic layer error: {env.redact(joined)}")
-        return body.get("data") or {}
+        return post_graphql(
+            url=self._url,
+            token=self._token,
+            timeout=self._timeout,
+            query=query,
+        )
 
     # ---- discovery ---------------------------------------------------------
-
-    # One document, one round trip, and every field on it verified against the
-    # live schema by introspection before it was written here. A field name the
-    # schema does not have fails the *whole* request, so the metrics list goes
-    # down with it: never add one speculatively. `label` on an entity is the
-    # standing example (the API answers "Cannot query field 'label' on type
-    # 'Entity'"), and `Measure` has no words either.
-    _CATALOG_FIELDS = (
-        "name type label description "
-        "queryableGranularities queryableTimeGranularities requiresMetricTime "
-        "filter { whereSqlTemplate } "
-        "typeParams { measure { name } inputMeasures { name } "
-        "numerator { name } denominator { name } expr "
-        "window { count granularity } grainToDate "
-        "metrics { name offsetWindow { count granularity } } } "
-        "measures { name agg expr aggTimeDimension } "
-        "semanticModels { name } "
-        "dimensions { name type label description expr semanticModel { name } "
-        "queryableGranularities queryableTimeGranularities } "
-        "entities { name type description expr role semanticModel { name } }"
-    )
 
     def list_definitions(self) -> SemanticCatalog:
         query = (
             "{ metrics(environmentId: "
             + self._env
             + ") { "
-            + self._CATALOG_FIELDS
+            + HOSTED_CATALOG_FIELDS
             + " } }"
         )
         data = self._post(query)
 
-        metrics: list[MetricInfo] = []
-        dims: dict[str, dict[str, Any]] = {}
-        roles: dict[str, dict[str, EntityRole]] = {}
-        entity_words: dict[str, str | None] = {}
-        measures: dict[str, MeasureInfo] = {}
-        models: set[str] = set()
-
-        for m in data.get("metrics") or []:
-            owners = [
-                s["name"]
-                for s in (m.get("semanticModels") or [])
-                if isinstance(s, dict) and s.get("name")
-            ]
-            models.update(owners)
-
-            for d in m.get("dimensions") or []:
-                owner = _model_name(d.get("semanticModel"))
-                if owner:
-                    models.add(owner)
-                bare = _bare_dimension(d.get("name"), owner)
-                merge_element_fields(
-                    dims,
-                    d.get("name"),
-                    {
-                        "type": d.get("type"),
-                        "label": d.get("label"),
-                        "description": d.get("description"),
-                        "definition": bare,
-                        "semantic_model": owner,
-                        "queryable_granularities": _grains(d),
-                        # Resolved against the bare name, never the queryable
-                        # token: a dimension that declares no `expr` references
-                        # the column its own name spells, and `user__pricing_tier`
-                        # is a path rather than a column anyone could select.
-                        "column": column_reference(d.get("expr"), bare),
-                    },
-                )
-
-            for e in m.get("entities") or []:
-                name, owner = e.get("name"), _model_name(e.get("semanticModel"))
-                if not name:
-                    continue
-                if owner:
-                    models.add(owner)
-                # Keyed by (entity, model) because that is the unit the layer
-                # declares: the same declaration is reached through every metric
-                # that can group by it, and two declarations of one entity
-                # genuinely differ in type, join key and description.
-                roles.setdefault(name, {}).setdefault(
-                    owner or "",
-                    EntityRole(
-                        semantic_model=owner or "",
-                        type=str(e.get("type") or "").lower(),
-                        expr=e.get("expr"),
-                        role=e.get("role"),
-                        description=e.get("description"),
-                        # Free: `expr` is already in the selection set, and the
-                        # rule that turns it into a column is the neutral one, so
-                        # a join key means the same thing on either backend even
-                        # though only one of them can name the relation it sits on.
-                        column=column_reference(e.get("expr"), name),
-                    ),
-                )
-                if entity_words.get(name) is None:
-                    entity_words[name] = e.get("description")
-
-            for measure in m.get("measures") or []:
-                name = measure.get("name")
-                if not name or name in measures:
-                    continue
-                measures[name] = MeasureInfo(
-                    name=name,
-                    agg=str(measure.get("agg") or "").lower() or None,
-                    expr=measure.get("expr"),
-                    agg_time_dimension=measure.get("aggTimeDimension"),
-                    column=column_reference(measure.get("expr"), name),
-                    # Deduced, not returned: the API carries no owning model on a
-                    # measure, and a measure lives in exactly one semantic model,
-                    # so a metric naming one model pins every measure it reads.
-                    # A measure reachable only through metrics that span several
-                    # models stays unset rather than being attributed to a guess.
-                    semantic_model=owners[0] if len(owners) == 1 else None,
-                )
-
-            metrics.append(_metric_info(m, owners))
-
-        notes = [_HOSTED_REACH_NOTE, _HOSTED_RELATION_NOTE]
-        if roles:
-            notes.append(_HOSTED_ENTITY_LABEL_NOTE)
-        disagreeing = sorted(m.name for m in metrics if len(m.time_axis or ()) > 1)
-        if disagreeing:
-            notes.append(
-                f"{', '.join(disagreeing)} aggregate over more than one time "
-                "column (see time_axis): grouping by metric_time uses each "
-                "measure's own, so the parts of one number can be bucketed by "
-                "different timestamps"
-            )
-        return SemanticCatalog.from_backend(
-            self,
-            semantic_models=[SemanticModelInfo(name=n) for n in sorted(models)],
-            metrics=metrics,
-            dimensions=[
-                DimensionInfo(
-                    name=name,
-                    type=fields.get("type") or "",
-                    label=fields.get("label"),
-                    description=fields.get("description"),
-                    definition=fields.get("definition"),
-                    semantic_model=fields.get("semantic_model"),
-                    queryable_granularities=fields.get("queryable_granularities"),
-                    column=fields.get("column"),
-                )
-                for name, fields in sorted(dims.items())
-            ],
-            entities=[
-                EntityInfo(
-                    name=name,
-                    type=derive_entity_type(list(declared.values())),
-                    description=entity_words.get(name),
-                    roles=[declared[key] for key in sorted(declared)],
-                )
-                for name, declared in sorted(roles.items())
-            ],
-            measures=[measures[name] for name in sorted(measures)],
-            notes=notes,
-        )
+        return decode_hosted_catalog(data, self)
 
     # ---- query -------------------------------------------------------------
 
@@ -701,7 +340,7 @@ class HostedDbtCloudBackend:
                     dimensions=[
                         d.get("name") for d in (m.get("dimensions") or []) if d
                     ],
-                    queryable_granularities=_grains(m),
+                    queryable_granularities=hosted_grains(m),
                 )
                 for m in (data.get("metrics") or [])
                 if m.get("name")
@@ -830,7 +469,7 @@ class HostedDbtCloudBackend:
         for alias, payload in data.items():
             if alias == "grains":
                 for metric in payload or []:
-                    grains = _grains(metric) if isinstance(metric, dict) else None
+                    grains = hosted_grains(metric) if isinstance(metric, dict) else None
                     # A metric the layer answered about, only. An unanswered one
                     # stays out of the map, which is what makes the requested grain
                     # pass through to the layer instead of being refused by dex.
@@ -918,50 +557,19 @@ class HostedDbtCloudBackend:
         return "[" + ", ".join(entries) + "]"
 
     def _await_result(self, query_id: str) -> Any:
-        for _ in range(self._POLL_ATTEMPTS):
-            query = (
-                f"{{ query(environmentId: {self._env}, queryId: "
-                f"{json.dumps(query_id)}) {{ status error "
-                f"jsonResult(encoded: false) }} }}"
-            )
-            data = self._post(query)
-            result = data.get("query") or {}
-            status = result.get("status")
-            if status == "SUCCESSFUL":
-                return result.get("jsonResult")
-            if status == "FAILED":
-                raise SemanticBackendError(
-                    "semantic layer query failed: "
-                    f"{env.redact(str(result.get('error')))}"
-                )
-            time.sleep(self._POLL_INTERVAL)
-        raise SemanticBackendError(
-            f"timed out waiting for semantic layer query {query_id}"
+        return await_result(
+            self._post,
+            environment_id=self._env,
+            query_id=query_id,
+            attempts=self._POLL_ATTEMPTS,
+            interval=self._POLL_INTERVAL,
         )
 
     def _shape(
         self, json_result: Any, *, extra_notes: list[str] | None = None
     ) -> dict[str, Any]:
-        payload = (
-            json.loads(json_result) if isinstance(json_result, str) else json_result
-        ) or {}
-        fields = (payload.get("schema") or {}).get("fields") or []
-        columns: list[str] = []
-        types: list[str] = []
-        for f in fields:
-            # `index` is the pandas row index, an artifact of orient='table'.
-            if f.get("name") == "index":
-                continue
-            columns.append(f.get("name"))
-            types.append(f.get("type"))
-        rows = payload.get("data") or []
-        cells = [[row.get(c) for c in columns] for row in rows]
-        return cap_columnar(
-            columns,
-            types,
-            cells,
-            max_rows=self._limits.max_rows,
-            max_cell_chars=self._limits.max_cell_chars,
-            max_payload_bytes=self._limits.max_payload_bytes,
+        return shape_json_result(
+            json_result,
+            limits=self._limits,
             extra_notes=extra_notes,
         )
